requirements.txt
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.3.0
streamlit==1.34.0
tensorflow==2.16.0rc0
opencv-python==4.9.0.80